refactor: 기존 헬퍼 클래스 리팩토링
