--- frontend/__tests__/playwright/helpers/test-helpers.ts
@@ -1,262 +1,0 @@
-import { Page, expect } from '@playwright/test';
-
-/**
- * 버튼 상호작용을 위한 베이스 헬퍼
- */
-export class BaseButtonHelper {
-  constructor(protected page: Page) {}
-
-  // 공통 버튼 상호작용 메서드들
-  private async clickButtonByText(name: string | RegExp) {
-    await this.page.getByRole('button', { name }).click();
-  }
-
-  private getButtonByText(name: string | RegExp) {
-    return this.page.getByRole('button', { name });
-  }
-
-  // 범용 헬퍼 (급할 때 직접 사용 가능)
-  async clickButton(name: string | RegExp) {
-    await this.clickButtonByText(name);
-  }
-
-  getButton(name: string | RegExp) {
-    return this.getButtonByText(name);
-  }
-}
-
-/**
- * 페이지 네비게이션 헬퍼 함수들
- */
-export class NavigationHelper {
-  constructor(private page: Page) {}
-
-  async goToHomePage() {
-    await this.page.goto('/');
-    await expect(this.page).toHaveTitle(/Coffee Shout|커피빵|2025-coffee-shout/i);
-  }
-
-  async waitForPageLoad(expectedUrl?: string | RegExp) {
-    await this.page.waitForLoadState('networkidle');
-    if (expectedUrl) {
-      await expect(this.page).toHaveURL(expectedUrl);
-    }
-  }
-}
-
-/**
- * 입력 폼 관련 헬퍼 함수들
- */
-export class FormHelper {
-  constructor(private page: Page) {}
-
-  async fillNickname(nickname: string) {
-    const nicknameInput = this.page.getByPlaceholder('닉네임을 입력해주세요');
-    await nicknameInput.fill(nickname);
-    await expect(nicknameInput).toHaveValue(nickname);
-  }
-
-  async fillJoinCode(joinCode: string) {
-    const joinCodeInput = this.page.getByPlaceholder('ex) ABCDE');
-    await joinCodeInput.fill(joinCode);
-    await expect(joinCodeInput).toHaveValue(joinCode);
-  }
-
-  async selectMenu(menuIndex: number = 0) {
-    // 셀렉트박스 클릭 (role="combobox" 사용)
-    await this.page.getByRole('combobox').click();
-
-    // 옵션 선택 (role="option" 사용)
-    const options = this.page.getByRole('option');
-    await options.nth(menuIndex).click();
-  }
-}
-
-/**
- * 의미있는 버튼 상호작용 헬퍼 함수들
- */
-export class ButtonHelper extends BaseButtonHelper {
-  // === 방 생성/참가 관련 ===
-  async clickCreateRoom() {
-    await this.clickButton('방 만들기');
-  }
-
-  async clickJoinRoomFromHome() {
-    await this.clickButton('방 참가하러 가기');
-  }
-
-  async clickJoinRoomFromModal() {
-    await this.clickButton('입장');
-  }
-
-  async clickGoToMenuSelection() {
-    await this.clickButton('메뉴 선택하러 가기');
-  }
-
-  async clickGoToCreateRoom() {
-    await this.clickButton('방 만들러 가기');
-  }
-
-  async clickEnterRoom() {
-    await this.clickButton('방 참가하기');
-  }
-
-  // === 로비 ===
-  async clickGameStart() {
-    await this.clickButton('게임 시작');
-  }
-
-  async clickGameReady() {
-    await this.clickButton('준비하기');
-  }
-
-  getGameWaitingButton() {
-    return this.getButton(/게임 대기중/);
-  }
-
-  getGameStartButton() {
-    return this.getButton('게임 시작');
-  }
-
-  getGameReadyButton() {
-    return this.getButton('준비하기');
-  }
-
-  getGameReadyCompleteButton() {
-    return this.getButton('준비 완료!');
-  }
-
-  async clickSectionTab(sectionName: '참가자' | '미니게임') {
-    await this.clickButton(sectionName);
-  }
-
-  async clickMiniGame(gameName: '카드게임') {
-    await this.clickButton(gameName);
-  }
-
-  // === 모달/공유 ===
-  async clickShare() {
-    await this.clickButton('공유');
-  }
-
-  async clickCopyIcon() {
-    await this.clickButton('초대 코드 복사');
-  }
-
-  async clickCloseModal() {
-    await this.clickButton('모달 닫기');
-  }
-
-  // === 룰렛 ===
-  async clickGoToRouletteStatus() {
-    await this.clickButton('룰렛 현황 보러가기');
-  }
-}
-
-/**
- * 페이지 검증 헬퍼 함수들
- */
-export class AssertionHelper {
-  constructor(private page: Page) {}
-
-  async expectToBeOnNicknamePage() {
-    await expect(this.page).toHaveURL(/\/entry\/name/);
-    await expect(this.page.getByPlaceholder('닉네임을 입력해주세요')).toBeVisible();
-  }
-
-  async expectToBeOnMenuSelectionPage() {
-    await expect(this.page).toHaveURL(/\/entry\/menu/);
-    await expect(this.page.getByRole('combobox')).toBeVisible();
-  }
-
-  async expectToBeOnLobbyPage() {
-    await expect(this.page).toHaveURL(/\/room\/.*\/lobby/);
-    await expect(this.page.getByRole('heading', { name: '참가자', level: 2 })).toBeVisible();
-  }
-
-  async expectToBeOnJoinCodePage() {
-    // 메인 페이지에서 모달이 열린 상태
-    await expect(this.page).toHaveURL('/');
-    await expect(this.page.getByPlaceholder('ex) ABCDE')).toBeVisible();
-  }
-
-  async expectShareModalToBeOpen() {
-    await expect(this.page.getByRole('dialog')).toBeVisible();
-    await expect(this.page.getByRole('heading', { name: '초대 코드' })).toBeVisible();
-  }
-}
-
-/**
- * 통합 테스트 헬퍼 클래스
- */
-export class TestHelper {
-  public navigation: NavigationHelper;
-  public form: FormHelper;
-  public button: ButtonHelper;
-  public assertion: AssertionHelper;
-
-  constructor(public page: Page) {
-    this.navigation = new NavigationHelper(page);
-    this.form = new FormHelper(page);
-    this.button = new ButtonHelper(page);
-    this.assertion = new AssertionHelper(page);
-  }
-
-  /**
-   * 호스트 방 생성 전체 플로우
-   */
-  async createRoomFlow(hostName: string, menuIndex: number = 0) {
-    await this.navigation.goToHomePage();
-    await this.button.clickCreateRoom();
-    await this.assertion.expectToBeOnNicknamePage();
-
-    await this.form.fillNickname(hostName);
-    await this.button.clickGoToMenuSelection();
-    await this.assertion.expectToBeOnMenuSelectionPage();
-
-    await this.form.selectMenu(menuIndex);
-    await this.button.clickGoToCreateRoom();
-    await this.assertion.expectToBeOnLobbyPage();
-  }
-
-  /**
-   * 게스트 방 참가 전체 플로우
-   */
-  async joinRoomFlow(joinCode: string, guestName: string, menuIndex: number = 0) {
-    await this.navigation.goToHomePage();
-    await this.button.clickJoinRoomFromHome();
-    await this.assertion.expectToBeOnJoinCodePage();
-
-    await this.form.fillJoinCode(joinCode);
-    await this.button.clickJoinRoomFromModal();
-    await this.assertion.expectToBeOnNicknamePage();
-
-    await this.form.fillNickname(guestName);
-    await this.button.clickGoToMenuSelection();
-    await this.assertion.expectToBeOnMenuSelectionPage();
-
-    await this.form.selectMenu(menuIndex);
-    await this.button.clickEnterRoom();
-    await this.assertion.expectToBeOnLobbyPage();
-  }
-
-  /**
-   * 호스트가 방 생성 후 초대 코드 가져오기
-   */
-  async createRoomAndGetJoinCode(hostName: string, menuIndex: number = 0): Promise<string> {
-    await this.createRoomFlow(hostName, menuIndex);
-    await this.button.clickShare();
-    await this.page.waitForSelector('h4');
-
-    const joinCodeElement = this.page.getByRole('heading', { level: 4 }).last();
-    const code = await joinCodeElement.textContent();
-
-    if (!code) {
-      throw new Error('방 참가 코드를 찾을 수 없습니다.');
-    }
-
-    await this.button.clickCloseModal();
-
-    return code.trim();
-  }
-}
